Extracts common libs and types
