+++ packages/lib/auth.ts
@@ -1,0 +1,11 @@
+import { compare, hash } from "bcryptjs";
+
+export async function hashPassword(password: string) {
+  const hashedPassword = await hash(password, 12);
+  return hashedPassword;
+}
+
+export async function verifyPassword(password: string, hashedPassword: string) {
+  const isValid = await compare(password, hashedPassword);
+  return isValid;
+}
