+++ packages/lib/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@calcom/lib",
+  "version": "0.0.0",
+  "main": "./index.ts",
+  "types": "./index.ts",
+  "license": "MIT",
+  "dependencies": {
+    "bcryptjs": "^2.4.3",
+    "dayjs": "^1.10.6",
+    "dayjs-business-time": "^1.0.4"
+  },
+  "devDependencies": {
+    "@calcom/tsconfig": "*",
+    "typescript": "^4.5.3"
+  }
+}
