+++ packages/lib/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@calcom/tsconfig/base.json",
+  "include": ["."],
+  "exclude": ["dist", "build", "node_modules"]
+}
