+++ packages/types/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@calcom/types",
+  "version": "0.0.0",
+  "private": true,
+  "main": "index.js"
+}
