+++ turbo.json
@@ -49,6 +49,9 @@
       ],
       "outputs": [".next/**"]
     },
+    "@calcom/web#dx": {
+      "dependsOn": ["@calcom/prisma#dx"]
+    },
     "@calcom/web#start": {},
     "build": {
       "dependsOn": ["^build"],
